Switched from set to vector for `PhysicalObject::Objects`
Because the order of `PhysicalObject::Objects` stored as `std::set< PhysicalObject  *>` depends on the pointer, worlds would store objects in an order that is not completely determined by the order of  insertion, and could change between different executions of the same code. Therefore, to ensure reproducibility, we switched to `std::vector<PhysicalObject  *>` which is orderer by insertion instead.